refactor: rename clients to providers
