+++ internal/whisper/providers/awssm/client.go
