+++ internal/whisper/providers/azkv/client_test.go
@@ -1,20 +1,20 @@
-package awssm_test
+package azkv_test
 
 import (
 	"context"
 	"fmt"
 	"log"
 
-	"github.com/busser/whisper/internal/whisper/clients/awssm"
+	"github.com/busser/whisper/internal/whisper/providers/azkv"
 )
 
 func Example() {
-	c, err := awssm.New()
+	c, err := azkv.New()
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	ref := "secret-sauce"
+	ref := "example.vault.azure.net/secret-sauce"
 	val, err := c.Resolve(context.Background(), ref)
 	if err != nil {
 		log.Fatal(err)
