+++ internal/whisper/providers/mock/client.go
