+++ internal/whisper/providers/passthrough/client.go
