+++ internal/whisper/providers/passthrough/client_test.go
